Create The Priest Mathematician.py
Problem Description:
One of the priests at the Benares temple (who loved the mathematics) assured their colleagues to achieve the transfer in the afternoon (the rhythm they had thought was a disc-per-second) by using an additional needle. They couldn’t believe him, but he proposed them the following strategy: 
• First move the topmost discs (say the top k discs) to one of the spare needles. 
• Then use the standard three needles strategy to move the remaining n − k discs (for a general case with n discs) to their destination. 
• Finally, move the top k discs into their final destination using the four needles. 
He calculated the value to k in order to minimize the number of movements and get a total of 18433 transfers, so they spent just 5 hours, 7 minutes and 13 seconds against the more than 500000 million years without the additional needle (as they would have to do 264 − 1 disc transfers. Can you believe it?)